[rabbitmq] Switch to March Hare 2.0
March Hare is the new name for Hot Bunnies.

2.0 release notes:
http://blog.rubyrabbitmq.info/blog/2013/12/05/march-hare-2-dot-0-0-is-released/

Logstash RabbitMQ plugins are not affected by the breaking API changes.
