+++ lib/logstash/inputs/rabbitmq.rb
@@ -11,6 +11,7 @@ require "logstash/namespace"
 # find links to both here:
 #
 # * RabbitMQ - <http://www.rabbitmq.com/>
+# * March Hare: <http://rubymarchhare.info>
 # * Bunny - <https://github.com/ruby-amqp/bunny>
 class LogStash::Inputs::RabbitMQ < LogStash::Inputs::Threadable
 
@@ -106,17 +107,17 @@ class LogStash::Inputs::RabbitMQ < LogStash::Inputs::Threadable
     super
   end
 
-  # Use HotBunnies on JRuby to avoid IO#select CPU spikes
+  # Use March Hare on JRuby to avoid IO#select CPU spikes
   # (see github.com/ruby-amqp/bunny/issues/95).
   #
-  # On MRI, use Bunny 0.9.
+  # On MRI, use Bunny.
   #
-  # See http://rubybunny.info and http://hotbunnies.info
+  # See http://rubybunny.info and http://rubymarchhare.info
   # for the docs.
   if RUBY_ENGINE == "jruby"
-    require "logstash/inputs/rabbitmq/hot_bunnies"
+    require "logstash/inputs/rabbitmq/march_hare"
 
-    include HotBunniesImpl
+    include MarchHareImpl
   else
     require "logstash/inputs/rabbitmq/bunny"
